# environment-a/api-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        example: unauthorized
        type: string
      status:
        example: error
        type: string
    type: object
  dto.FileDTO:
    properties:
      chunk_count:
        type: integer
      created_at:
        type: string
      dedup_ratio:
        type: number
      deleted_at:
        type: string
      directory_id:
        type: string
      id:
        type: string
      manifest_id:
        type: string
      mime_type:
        type: string
      name:
        type: string
      new_chunk_count:
        type: integer
      reuse_chunk_count:
        type: integer
      size_bytes:
        type: integer
      starred_at:
        type: string
    type: object
  dto.FileDetailResponse:
    properties:
      file:
        $ref: '#/definitions/dto.FileDTO'
      status:
        type: string
    type: object
  dto.FileListResponse:
    properties:
      directory_id:
        type: string
      files:
        items:
          $ref: '#/definitions/dto.FileDTO'
        type: array
      status:
        type: string
    type: object
  dto.FileManifestResponse:
    properties:
      file:
        $ref: '#/definitions/dto.FileDTO'
      manifest:
        $ref: '#/definitions/dto.ManifestInfoDTO'
      status:
        type: string
    type: object
  dto.FileMutationResponse:
    properties:
      file:
        $ref: '#/definitions/dto.FileDTO'
      status:
        type: string
    type: object
  dto.FileSearchResponse:
    properties:
      directory_id:
        type: string
      files:
        items:
          $ref: '#/definitions/dto.FileDTO'
        type: array
      include_deleted:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      query:
        type: string
      status:
        type: string
      total:
        type: integer
    type: object
...
  dto.ManifestInfoDTO:
    properties:
      chunk_count:
        type: integer
      created_at:
        type: string
      file_hash:
        type: string
      immutable:
        type: boolean
      manifest_id:
        type: string
      total_size_bytes:
        type: integer
    type: object
...
  dto.OKResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
  dto.SoftDeleteResponse:
    properties:
      deleted_at:
        type: string
      file_id:
        type: string
      status:
        type: string
    type: object
...
  dto.UploadCommitResultDTO:
    properties:
      chunk_count:
        type: integer
      dedup_ratio:
        type: number
      file_hash:
        type: string
      immutable:
        type: boolean
      manifest_id:
        type: string
      new_chunk_count:
        type: integer
      reuse_chunk_count:
        type: integer
      total_size_bytes:
        type: integer
    type: object
  dto.UploadResponse:
    properties:
      file:
        $ref: '#/definitions/dto.FileDTO'
      status:
        type: string
      upload_commit_result:
        $ref: '#/definitions/dto.UploadCommitResultDTO'
    type: object
...
info:
  contact: {}
  description: API service for authentication, directory hierarchy, and immutable
    file metadata.
  title: HashBox API Service
  version: "1.0"
paths:
...
  /files:
    get:
      description: List root files or files inside a directory when directory_id is
        provided
      parameters:
      - description: Directory UUID; omit for root files
        in: query
        name: directory_id
        type: string
      - description: Include soft-deleted files
        in: query
        name: include_deleted
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FileListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List files
      tags:
      - files
    post:
      consumes:
      - multipart/form-data
      description: Stream multipart file upload to vault-core and persist metadata
      parameters:
      - description: Directory UUID; omit for root upload
        in: formData
        name: directory_id
        type: string
      - description: File payload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload file
      tags:
      - files
  /files/{id}:
    delete:
      description: Mark file as deleted without removing physical payload
      parameters:
      - description: File UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SoftDeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft delete file
      tags:
      - files
    get:
      description: Get file metadata by ID (including soft-deleted files)
      parameters:
      - description: File UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FileDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: File detail
      tags:
      - files
  /files/{id}/download:
    get:
      description: Stream file payload from vault-core by file ID
      parameters:
      - description: File UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download file
      tags:
      - files
  /files/{id}/manifest:
    get:
      description: Get immutable manifest details from vault-core for a file
      parameters:
      - description: File UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FileManifestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: File manifest info
      tags:
      - files
  /files/{id}/permanent:
    delete:
      description: Permanently remove deleted file metadata from trash
      parameters:
      - description: File UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Permanently delete file
      tags:
      - files
  /files/{id}/restore:
    post:
      description: Restore a previously deleted file
      parameters:
      - description: File UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FileMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore file from trash
      tags:
      - files
  /files/{id}/star:
    delete:
      description: Remove starred marker from a file
      parameters:
      - description: File UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FileMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unstar file
      tags:
      - files
    put:
      description: Mark a file as starred
      parameters:
      - description: File UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FileMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Star file
      tags:
      - files
  /files/search:
    get:
      description: Search file metadata by query in user namespace with optional filters
        and pagination
      parameters:
      - description: Search keyword (min 2 chars)
        in: query
        name: q
        required: true
        type: string
      - description: Directory UUID
        in: query
        name: directory_id
        type: string
      - description: Include soft-deleted files
        in: query
        name: include_deleted
        type: boolean
      - description: Page size (1-200, default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FileSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search files
      tags:
      - files
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"